Use Correct TFM versions of microsoft packages that follows dotnet versioning
The package Microsoft.AspNetCore.Authorization at Version 9.0.9 does not publish builds for net8.0
so apps default to using the version built for netstandard2.0. A new property AllowsCachingPolicies is
not included in the netstandard build, so apps running net8.0 with `Altinn.Common.PEP` crash at
runtime in the middleware, because of the version missmatch.

The suggested solution here is to ensure that `Altinn.Common.PEP` requires version `8.0.20` of
`Microsoft.AspNetCore.Authorization` when it is built for net8.0